moved movie details component to listings file
